refactor: split the whole-list tab lifecycle out of terminal-tabs-core
terminal-tabs-core.ts crossed the ~500-line cap. The seam it was hiding is
where a TabsState comes from: first mount, a restart snapshot, re-entering an
emptied task and the last tab exiting all mint a whole list from something
other than a user action, and they must agree about ordinal minting and
carried identity — which is exactly what recycle and reopen had stopped doing.
Those four move to terminal-tabs-lifecycle.ts (re-exported by core, so no
importer changes); TabsState joins the other shapes in terminal-tab-shapes.ts
so the new module needs no runtime import from core. The test file splits the
same way.
